clean up user tabs section of user page
